Subject: [PATCH 3/4] Device: reorganize endpoint decoders initialization
For some reason, that I failed to discover, the current way in which all
endpoint decoders are gathered for a device fails to work properly for
the UsbmonLogParser. (The Storage decoders never were used to decode the
relevant lines in a usbmon log for a USB Mass Storage device.)

A small reorganization of the Device class somehow solves this.
Basically, it makes sure all devices always use controlDecoder for
endpoint 0 and that new decoders are added to a device whenever it turns
out they are needed. That seems to be the logical thing to do anyway.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>



git-svn-id: https://svn.code.sf.net/p/vusb-analyzer/code/trunk@23 31ffde5e-7fae-4e65-b3f5-864f19086f7f