UsbmonLogParser: correctly parse iso transactions
The comments described the current output format (with room for one
'number of isochronous frame descriptors' token and zero or more
'descriptor' tokens). The code assumed neither were present (so it expected
usbmon's legacy format).

The code to parse the current output format for isochronous transactions
is rather messy, but it is hard to see how these could be parsed more
cleanly ... (Note that we don't actually parse the descriptors as the
information they provide isn't used anywhere.)

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>

---
 VUsbTools/Log.py |   30 ++++++++++++++++++++++++++----
 1 files changed, 26 insertions(+), 4 deletions(-)




git-svn-id: https://svn.code.sf.net/p/vusb-analyzer/code/trunk@29 31ffde5e-7fae-4e65-b3f5-864f19086f7f